Extract PipelineContext from RuleAlteredContext
